Instruction rewritten to TS
