--- src/components/Instruction/Instruction.js
@@ -1,52 +1,0 @@
-import StepNumber from '@components/StepNumber'
-import { useStyles } from '@helpers/hooks/useStyles'
-import PropTypes from 'prop-types'
-import React from 'react'
-import uuid from 'react-uuid'
-
-import styles from './Instruction.module.scss'
-
-/**
- * Instruction - stateless presentational component
- * @param {object} props - props
- * @param {string|array} props.className - a list of class names, default: `instruction`
- * @param {array} props.data - steps data
- * @return {object} An object of children element
- */
-function Instruction({ className, data }) {
-  const instructionStyles = useStyles(
-    {
-      [styles.instruction]: true
-    },
-    className
-  )
-
-  return (
-    <div className={instructionStyles}>
-      {data.map(({ content }, index) => (
-        <div className={styles.instruction__step} key={uuid()}>
-          <StepNumber absolute size='medium' step={index + 1} />
-
-          {content}
-        </div>
-      ))}
-    </div>
-  )
-}
-
-Instruction.displayName = 'Instruction'
-
-Instruction.propTypes = {
-  className: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
-  data: PropTypes.arrayOf(
-    PropTypes.shape({
-      content: PropTypes.node.isRequired
-    })
-  ).isRequired
-}
-
-Instruction.defaultProps = {
-  className: ''
-}
-
-export default Instruction
