+++ src/components/Instruction/Instruction.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Instruction from '@components/Instruction'
+import { Instruction } from '@components/Instruction'
 
 # Instruction
