+++ src/components/Instruction/index.ts
@@ -1,0 +1,1 @@
+export * from './Instruction'
