+++ src/components/index.ts
@@ -45,7 +45,7 @@ export { default as InfoBar } from './InfoBar'
 export * from './InfoSection'
 export { default as Input } from './Input'
 export { default as MaskedInput } from './Input/MaskedInput'
-export { default as Instruction } from './Instruction'
+export * from './Instruction'
 export { default as Label } from './Label'
 export { default as Layout } from './Layout'
 export { default as Legend } from './Legend'
